Dev (#49)
* refactor: fixed and refactored path, plots, and small QoL bugs in preperation for release (first version)

* fixes: plots now optionally follow, and can be scaled according to user choice

* fixes

* fixes: correct added postbuild copmmand for HDF5 dlls, along with 3rd party licenses in new release

* update dev inline with Main branch (#42)

* Update citation link in README.md

* Update author information and release date (#41)

* alpha release v0.1.0 (#40)

* refactor: fixed and refactored path, plots, and small QoL bugs in preperation for release (first version)

* fixes: plots now optionally follow, and can be scaled according to user choice

* fixes

* fixes: correct added postbuild copmmand for HDF5 dlls, along with 3rd party licenses in new release

* Remove badge section from README.md

Removed badge section from README.

* fixes

* Feature/joint friction remodel (#44)

* feat: methods to compute linear velocity and angular velocity jacobians

* WIP: addition of jacobians, and replacement of glm with Eigen in mathematically dense compuations

* refactor: move eigen-based mathematical computations for non-visual computational functions.

* feat: added mass matrix logic to M(joint, link)

* fixes: physics computations surrounding model-numeric mistmatch fixes

* feat: Coriolis Model added

NOTE:
- Does not work properly, may crash program due to high output of computations

* fixes: corrected logic and pipeline for inertia and coriolis computations

NOTES:
- Replace computeI_eff with computeSingleIeff for modular, per-joint inertia calculation.
- Rewrite computeCoriolisDiagonal to use finite differences and new inertia method.
- Simplify deriv() to compute inertia and Coriolis terms internally from state.
- Improve computeJointInertiaContribution to use both translational and rotational effects with joint axis in world frame.
- Update simulation and integration logic to u
- I did a final check with github copilot as this is quite abstract for me

* Refactor/robot metrics (#46)

* refactor: update metrics output to suit report question, and limit file size of each run.

* feat: added save command for log file naming and saving instead of always outputting.

TODO:
- Implement gravity torque into current joint-space formula.
- Add different arms to showcase different abilities.
- Add a general script that works for multiple arms.
- Fix RotateX commands (not essential).
- Get save working for specific file names (not essential).
- Use either python or MATLAB to generate plots for the run data.

NOTES:
- I am using various resources to improve this, new ground for me.

* refactor: undid ALOT of thread work, not necessary and way too time consuming.

READY FOR DATA ANALYSIS

* Update Sim_Engine/EngineLib/include/Robots/RobotSystem.h


* Update Sim_Engine/EngineLib/include/Robots/RobotModel.h


* Update MathLib/MathLib/src/integrators/numerical_integrators.cpp


* Update Sim_Engine/EngineLib/include/Interpreter/Commands/SaveCmd.h


* Update Sim_Engine/EngineLib/include/Interpreter/Commands/SaveCmd.h


* Update Sim_Engine/EngineLib/src/Interpreter/Parser.cpp


* Update Sim_Engine/EngineLib/src/Numerics/IntegrationService.cpp


---------


* feat: implemented basic computation of a joints gravity torque,

NOTES:
- Removed the hardcoded setting of tau_i to 0

* fixes: fixed computational overestimation of wrist gravity in ComputeGravityTorque()

* fixes: fixed gravity inconsistenties with robot root frame, and recomputation of FK

NOTES:
- This is confusing, but also not, its very much not in my comfort zone

* fixes: Inverse dynamics style feedforward corrected to be inline with the physically correct model.

* feat: added a very basic selector for gravity values

* fixes

* fixes

* merge fixes

* merge fixes

* merge fixes

* merge fixes

---------

+++ MathLib/MathLib/src/integrators/numerical_integrators.cpp
@@ -65,42 +65,42 @@ namespace integration {
 			const double c7 = 1.0;
 
 			// Dormand-Prince coefficients
-			const double a21 = 1.0 / 5.0;
-			const double a31 = 3.0 / 40.0;
-			const double a32 = 9.0 / 40.0;
-			const double a41 = 44.0 / 45.0;
+			const double a21 =  1.0 / 5.0;
+			const double a31 =  3.0 / 40.0;
+			const double a32 =  9.0 / 40.0;
+			const double a41 =  44.0 / 45.0;
 			const double a42 = -56.0 / 15.0;
-			const double a43 = 32.0 / 9.0;
-			const double a51 = 19372.0 / 6561.0;
+			const double a43 =  32.0 / 9.0;
+			const double a51 =  19372.0 / 6561.0;
 			const double a52 = -25360.0 / 2187.0;
-			const double a53 = 64448.0 / 6561.0;
+			const double a53 =  64448.0 / 6561.0;
 			const double a54 = -212.0 / 729.0;
-			const double a61 = 9017.0 / 3168.0;
+			const double a61 =  9017.0 / 3168.0;
 			const double a62 = -355.0 / 33.0;
-			const double a63 = 46732.0 / 5247.0;
-			const double a64 = 49.0 / 176.0;
+			const double a63 =  46732.0 / 5247.0;
+			const double a64 =  49.0 / 176.0;
 			const double a65 = -5103.0 / 18656.0;
-			const double a71 = 35.0 / 384.0;
-			const double a72 = 0.0;
-			const double a73 = 500.0 / 1113.0;
-			const double a74 = 125.0 / 192.0;
+			const double a71 =  35.0 / 384.0;
+			const double a72 =  0.0;
+			const double a73 =  500.0 / 1113.0;
+			const double a74 =  125.0 / 192.0;
 			const double a75 = -2187.0 / 6784.0;
-			const double a76 = 11.0 / 84.0;
+			const double a76 =  11.0 / 84.0;
 
 			// Weights for 4th and 5th order estimates
-			const double b1 = 35.0 / 384.0;
-			const double b2 = 0.0;
-			const double b3 = 500.0 / 1113.0;
-			const double b4 = 125.0 / 192.0;
-			const double b5 = -2187.0 / 6784.0;
-			const double b6 = 11.0 / 84.0;
-			const double b1s = 5179.0 / 57600.0;
-			const double b2s = 0.0;
-			const double b3s = 7571.0 / 16695.0;
-			const double b4s = 393.0 / 640.0;
+			const double b1  =  35.0 / 384.0;
+			const double b2  =  0.0;
+			const double b3  =  500.0 / 1113.0;
+			const double b4  =  125.0 / 192.0;
+			const double b5  = -2187.0 / 6784.0;
+			const double b6  =  11.0 / 84.0;
+			const double b1s =  5179.0 / 57600.0;
+			const double b2s =  0.0;
+			const double b3s =  7571.0 / 16695.0;
+			const double b4s =  393.0 / 640.0;
 			const double b5s = -92097.0 / 339200.0;
-			const double b6s = 187.0 / 2100.0;
-			const double b7s = 1.0 / 40.0;
+			const double b6s =  187.0 / 2100.0;
+			const double b7s =  1.0 / 40.0;
 
 			// Compute the Runge-Kutta stages (DP -> 7 stages)
 			const VecX k1 = f(t, x);
@@ -144,7 +144,7 @@ namespace integration {
 			}
 		}
 
-		// If we reach here, it means we failed to converge after many attempts
+		// If it reaches here, it failed to converge after many attempts
 		// Just putting this here as a policy choice honestly, return the best effort or throw
 		throw std::runtime_error("RK45 failed to converge after maximum attempts");
 		//return x;
@@ -154,7 +154,7 @@ namespace integration {
 	// ----------------------------------------------------------------
 	// Finite Difference Method
 	VecX PDE::fdmStep(const VecX& u, double dx, double dt, double alpha) {
-		int n = u.size();
+		int n = (int)u.size();
 		VecX u_new = u;
 		double r = alpha * dt / (dx * dx);
 		for (int i = 1; i < n - 1; ++i) {
